Update function lib
