+++ nodejs/llama.js
@@ -2,12 +2,12 @@ require("dotenv").config();
 const { Web3 } = require("web3");
 const ethers = require("ethers");
 const concat = require("concat-stream");
-const network = require("./network.json");
-const web3 = new Web3(process.env.FROM_TESTNET_RPC);
-const oaocabi = require("./build/contracts/OAOCircle.json")["abi"];
+const network = require("../network.json");
+const web3 = new Web3(process.env.DEFAULT_TESTNET_RPC);
+const oaocabi = require("../build/contracts/OAOCircle.json")["abi"];
 
 const SENDER = web3.eth.accounts.privateKeyToAccount(
-  process.env.DEFAULT_PRIVATE_KEY,
+  process.env.FROM_PRIVATE_KEY,
 );
 web3.eth.accounts.wallet.add(SENDER);
 const OAOC_ADDRESS = "0xf59eBE57B47c51f8583264be7e21692fCB211AB4"; // OAOCircle
@@ -16,7 +16,7 @@ const OAOCircle = new web3.eth.Contract(oaocabi, OAOC_ADDRESS, {
   from: SENDER.address,
 });
 const TO = 3600 * 1000; // Due to LLama inference
-const MSG_VALUE = 0.02;
+const MSG_VALUE = 0.1;
 
 async function waitForTransaction(web3, txHash) {
   let transactionReceipt = await web3.eth.getTransactionReceipt(txHash);
@@ -64,7 +64,7 @@ async function oao(_instruction) {
     "Answer about the transaction detail.\n" +
     "Give me totaly only five words answer!\n";
   // condition = ""
-  var prompt = condition.concat(_instruction);
+  var prompt = await condition.concat(_instruction);
   console.log(`prompt:\n${prompt}`);
   const tx = await OAOCircle.methods
     .calculateAIResult(prompt)
@@ -76,4 +76,4 @@ async function oao(_instruction) {
   return await parseResult(result);
 }
 
-main();
+module.exports = { oao }
